# apps/web/docs/openapi.yaml
openapi: 3.0.3
info:
  title: QuizForge API
  version: 1.0.0
servers:
  - url: http://localhost:4000
paths:
...
  /exams/{id}:
    get:
      summary: Get exam
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Exam }
    patch:
      summary: Update exam
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Exam }
...